# services/user/docs/swagger.yaml
basePath: /api/v1/user
definitions:
  errorx.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 音频 AI 平台用户服务接口文档，包含用户登录、注册、个人信息、家庭管理、设备分享、下载管理、实名认证等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 用户服务 API 文档
  version: "1.0"
paths:
...
  /user/cancellation/withdraw:
    post:
      consumes:
      - application/json
      description: 在冷静期内撤销注销申请，恢复账号
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 撤销注销
      tags:
      - 账号注销
...
  /user/download/resume:
    post:
      consumes:
      - application/json
      description: 中断后继续下载，返回断点位置
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 继续下载
      tags:
      - 下载服务
...
  /user/family/member/role/update:
    post:
      consumes:
      - application/json
      description: 修改家庭成员的角色
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 更新成员角色
      tags:
      - 家庭管理
  /user/info:
    put:
      consumes:
      - application/json
      description: 已登录用户更新个人资料，如昵称、头像等
      parameters:
      - description: 用户信息更新请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 更新用户资料
      tags:
      - 用户信息
...
  /user/realname/submit:
    post:
      consumes:
      - application/json
      description: 提交实名认证申请，等待后台审核
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 提交实名信息
      tags:
      - 实名认证
...
  /user/verify/send:
    post:
      consumes:
      - application/json
      description: 发送验证码到邮箱或手机，用于注册、登录、找回密码等场景
      parameters:
      - description: 发送验证码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "429":
          description: 发送过于频繁
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      summary: 发送验证码
      tags:
      - 验证码
securityDefinitions:
  BearerAuth:
    description: 请输入 Bearer Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"