# services/user/internal/handler/update_user_info_handler.go
// Code scaffolded by goctl. Safe to edit.
// goctl 1.9.2

package handler

import (
	"net/http"

	"github.com/jacklau/audio-ai-platform/services/user/internal/logic"
	"github.com/jacklau/audio-ai-platform/services/user/internal/svc"
	"github.com/jacklau/audio-ai-platform/services/user/internal/types"
	"github.com/zeromicro/go-zero/rest/httpx"
)

// UpdateUserInfoHandler 更新用户资料处理器
// @Summary      更新用户资料
// @Description  已登录用户更新个人资料，如昵称、头像等
// @Tags         用户信息
// @Accept       json
// @Produce      json
// @Param        body  body      errorx.Response  true  "用户信息更新请求"
// @Success      200  {object}  errorx.Response  "更新成功"
// @Failure      400  {object}  errorx.Response  "参数错误"
// @Failure      401  {object}  errorx.Response  "未登录"
// @Failure      500  {object}  errorx.Response  "服务器错误"
// @Router       /user/info [put]
// @Security     BearerAuth
// POST /api/v1/user/info/update
// 用途：已登录用户更新个人资料，JWT 定身份，body 传要改的字段
func UpdateUserInfoHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.UpdateUserInfoReq
		if err := httpx.Parse(r, &req); err != nil {
			httpx.ErrorCtx(r.Context(), w, err)
			return
		}

		l := logic.NewUpdateUserInfoLogic(r.Context(), svcCtx)
		resp, err := l.UpdateUserInfo(&req)
		if err != nil {
			httpx.ErrorCtx(r.Context(), w, err)
		} else {
			httpx.OkJsonCtx(r.Context(), w, resp)
		}
	}
}